Split symtable implementation
